pyproject.toml
[tool.poetry]
name = "python-clean-architecture"
version = "0.1.0"
description = ""
authors = ["Vinicius <vinigmoraes@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "2.0.1"
Flask-RESTful = "0.3.9"
flask-expects-json = "^1.6.0"
Flask-Injector = "0.12.3"
dataclasses-json = "^0.5.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
